source/requirements.txt
Pandas
polars
pyarrow
requests
fastapi
uvicorn
pydantic
tqdm
python-dotenv
pyyaml